requirements-dev.txt
types-requests==2.32.0.20250328
types-PyYAML==6.0.12.20250402
mypy-extensions==1.0.0
mypy==1.15.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework-stubs==3.15.3
dockerfile==3.4.0
psycopg2==3.2.9
файлы которые не нужны в докере